Fixed the allocation of SDSC tags within Master System ROM images.
Previously this would try to allocate starting from address $0000 on the page containing range $7FE0..$7FFF, resulting in it thinking address $0000 was free. With a large 32KB this worked fine, but if using two 16KB pages the start address of the page containing $7FE0..$7FFF is $4000, not $0000, so that's where the search for free space should go.